feat: add image variation api
